+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All Notable changes to `flipboxdigital/queue` will be documented in this file
 
+## Unreleased
+### Changed
+- When using multiple queues, passing a null queue index will result in getting the first registered queue.
+
 ## 1.0.0-beta.1 - 2017-6-15
 ### Added
 - `JobInterface::toConfig() is responsible for packaging the job for serialization.`